Big update reformat

+++ src/main/java/spring_boot/project_swp/repository/RoleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import spring_boot.project_swp.entity.Role;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> { // <--- SỬA Integer THÀNH Long
   Optional<Role> findByRoleName(String roleName);
 }
